synthesizer/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   groq
    #   httpx
    #   starlette
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
click==8.1.8
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   loguru
distro==1.9.0
    # via groq
dotenv==0.9.9
    # via synthesizer (pyproject.toml)
fastapi==0.115.11
    # via synthesizer (pyproject.toml)
groq==0.22.0
    # via synthesizer (pyproject.toml)
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   groq
    #   ollama
idna==3.10
    # via
    #   anyio
    #   httpx
kafka-python==2.1.2
    # via synthesizer (pyproject.toml)
loguru==0.7.3
    # via synthesizer (pyproject.toml)
ollama==0.4.7
    # via synthesizer (pyproject.toml)
pydantic==2.10.6
    # via
    #   fastapi
    #   groq
    #   ollama
pydantic-core==2.27.2
    # via pydantic
python-dotenv==1.0.1
    # via dotenv
redis==5.2.1
    # via synthesizer (pyproject.toml)
sniffio==1.3.1
    # via
    #   anyio
    #   groq
starlette==0.46.1
    # via fastapi
typing-extensions==4.12.2
    # via
    #   fastapi
    #   groq
    #   pydantic
    #   pydantic-core
uvicorn==0.34.0
    # via synthesizer (pyproject.toml)
win32-setctime==1.2.0
    # via loguru
